refactor: tobe package의 view, annotationmapping package 분리
