+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/handlermapping/annotationmapping/HandlerKey.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.handlermapping.annotationmapping;
 
 import nextstep.web.annotation.RequestMethod;
