+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/AnnotationHandlerMappingTest.java
@@ -1,7 +1,8 @@
 package nextstep.mvc.tobe;
 
 import nextstep.db.DataBase;
-import nextstep.mvc.tobe.handlermapping.AnnotationHandlerMapping;
+import nextstep.mvc.tobe.handlermapping.annotationmapping.AnnotationHandlerMapping;
+import nextstep.mvc.tobe.handlermapping.annotationmapping.HandlerExecution;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
