+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/HandlerKeyTest.java
@@ -1,5 +1,6 @@
 package nextstep.mvc.tobe;
 
+import nextstep.mvc.tobe.handlermapping.annotationmapping.HandlerKey;
 import nextstep.web.annotation.RequestMethod;
 import org.junit.jupiter.api.Test;
