+++ slipp/src/main/java/slipp/SlippWebApplicationInitializer.java
@@ -1,7 +1,7 @@
 package slipp;
 
 import nextstep.mvc.DispatcherServlet;
-import nextstep.mvc.tobe.handlermapping.AnnotationHandlerMapping;
+import nextstep.mvc.tobe.handlermapping.annotationmapping.AnnotationHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
